moving SmallVector.h
