+++ include/cudaq/Optimizer/Dialect/CC/CCTypes.h
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include "cudaq/Optimizer/Dialect/Common/SmallVector.h"
+#include "cudaq/Support/SmallVector.h"
 #include "mlir/IR/BuiltinAttributes.h"
 #include "mlir/IR/BuiltinTypes.h"
 #include "mlir/IR/Types.h"
